#160 added an associated entity component that holds another entity (usually the gameobject)

+++ engine/src/engine/external/actions/DirectionAction.java
@@ -3,7 +3,7 @@ package engine.external.actions;
 import engine.external.component.DirectionComponent;
 
 public class DirectionAction extends NumericAction {
-    public DirectionAction(ModifyType type, Integer direction) {
+    public DirectionAction(ModifyType type, Double direction) {
         setAction(type, direction, DirectionComponent.class);
     }
 }
